refactored and working strategy pattern
