+++ apps/notification/src/common/errors/dto-validation.error.ts
@@ -1,0 +1,1 @@
+export class DtoValidationException extends Error {}
