+++ apps/notification/src/notification/errors/notifier-strategy.error.ts
@@ -1,0 +1,9 @@
+import { DeliveryMethods } from '@hermes/common';
+
+export class NotifierStrategyException extends Error {
+  constructor(type: DeliveryMethods) {
+    super(
+      `Invalid Delivery Method: ${type} does not have a defined NotifierStrategy`,
+    );
+  }
+}
