+++ apps/notification/src/notification/interfaces/notifier-strategy.interface.ts
@@ -1,8 +1,6 @@
-import { DeliveryMethods } from '@hermes/common';
 import { NotificationDto } from './notification-dto.interface';
 
 export interface NotifierStrategy<T extends NotificationDto> {
-  type: DeliveryMethods;
   createNotificationDto(data: any): Promise<T>;
   createTemplate(dto: T): Promise<T>;
   notify(dto: T): Promise<any>;
